Pass ValueInRange to NormalizedMurmurHash
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3968985
Commit-Queue: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062332}

+++ components/variations/variations_test_utils.h
@@ -130,7 +130,7 @@ class MockEntropyProviders : public EntropyProviders {
     absl::optional<double> high_entropy = absl::nullopt;
   };
   explicit MockEntropyProviders(Results results,
-                                size_t low_entropy_domain = 8000);
+                                uint32_t low_entropy_domain = 8000);
   ~MockEntropyProviders() override;
 
   const base::FieldTrial::EntropyProvider& low_entropy() const override;
